// src/types/responses.ts
export interface BibleSummary {
  id: string;
  dblId: string;
  abbreviation: string;
  abbreviationLocal: string;
  language: Language;
  countries: {
    id: string;
    name: string;
    nameLocal: string;
  }[];
  name: string;
  nameLocal: string;
  description: string;
  descriptionLocal: string;
  relatedDbl: string;
  type: string;
  /** @format date-time */
  updatedAt: string;
  audioBibles?: AudioBibleSummary[];
}

export interface AudioBibleSummary {
  id: string;
  name: string;
  nameLocal: string;
  description: string;
  descriptionLocal: string;
}
// ...
export interface Book {
  id: string;
  bibleId: string;
  abbreviation: string;
  name: string;
  nameLong: string;
  chapters?: ChapterSummary[];
}

export interface ChapterSummary {
  id: string;
  bibleId: string;
  number: string;
  bookId: string;
  reference: string;
}

export interface Chapter {
  id: string;
  bibleId: string;
  number: string;
  bookId: string;
  content: string;
  reference: string;
  verseCount: number;
  next?: {
    id?: string;
    bookId?: string;
    number?: string;
  };
  previous?: {
    id?: string;
    bookId?: string;
    number?: string;
  };
  copyright: string;
}

export interface Language {
  id: string;
  name: string;
  nameLocal: string;
  script: string;
  scriptDirection: string;
}

export interface Passage {
  id: string;
  bibleId: string;
  orgId: string;
  content: string;
  reference: string;
  verseCount: number;
  copyright: string;
}
// ...
export interface SearchVerse {
  id: string;
  orgId: string;
  bibleId: string;
  bookId: string;
  chapterId: string;
  text: string;
  reference?: string;
}

export interface SearchResponse {
  query: string;
  limit: number;
  offset: number;
  total: number;
  verseCount: number;
  verses: SearchVerse[];
  passages?: Passage[];
}